refactor(silencer): split control-queue glue into headless.cpp
Phase 4 mechanical move. DrainControlQueue and PostFrameReplies now
live in src/game/headless.cpp. Declarations stay on Game in game.h.
No behavior change.
